requirements.txt
aiogram==2.6.1
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
idna==2.8
multidict==4.7.4
pycparser==2.19
pyTelegramBotAPI==3.6.7
python-telegram-bot==12.3.0
pytz==2019.3
requests==2.22.0
six==1.14.0
tornado==6.0.3
urllib3==1.25.8
yarl==1.4.2
